requirements.txt
pytest>=7.2.2
pytest-xdist
PyYAML>=6.0
gin-config
absl-py>=1.2.0
GPUtil>=1.4.0
tqdm>=4.64.1
sox

numpy==2.2.5
scipy==1.15.2
librosa>=0.9.2
scikit_learn>=1.1.2

torch>=2.5.1
torchaudio>=2.5.1
tensorboard
pytorch_lightning
torch_ema
einops>=0.5.0
einops_exts>0.0.4

nn-tilde>=1.5.2
cached-conv>=2.5.0
Flask>=2.2.3
